+++ src/implementations/CollarVaultManager.sol
@@ -30,11 +30,11 @@ contract CollarVaultManager is ERC6909TokenSupply, Ownable, ICollarVaultManager 
 
     // ----- STATE VARIABLES ----- //
 
-    uint256 public vaultCount;
+    uint public vaultCount;
 
     mapping(bytes32 uuid => Vault vault) internal vaultsByUUID;
-    mapping(uint256 vaultNonce => bytes32 UUID) public vaultsByNonce;
-    mapping(bytes32 => uint256) public vaultTokenCashSupply;
+    mapping(uint vaultNonce => bytes32 UUID) public vaultsByNonce;
+    mapping(bytes32 => uint) public vaultTokenCashSupply;
 
     // ----- CONSTRUCTOR ----- //
 
@@ -58,7 +58,7 @@ contract CollarVaultManager is ERC6909TokenSupply, Ownable, ICollarVaultManager 
         return abi.encode(vaultsByUUID[uuid]);
     }
 
-    function vaultInfoByNonce(uint256 vaultNonce) external view override returns (bytes memory) {
+    function vaultInfoByNonce(uint vaultNonce) external view override returns (bytes memory) {
         bytes32 uuid = vaultsByNonce[vaultNonce];
         if (vaultsByUUID[uuid].openedAt == 0) {
             revert InvalidVault();
@@ -67,19 +67,11 @@ contract CollarVaultManager is ERC6909TokenSupply, Ownable, ICollarVaultManager 
         return abi.encode(vaultsByUUID[uuid]);
     }
 
-    function getVaultUUID(uint256 vaultNonce) external view override returns (bytes32 uuid) {
+    function getVaultUUID(uint vaultNonce) external view override returns (bytes32 uuid) {
         return vaultsByNonce[vaultNonce];
     }
 
-    function previewRedeem(
-        bytes32 uuid,
-        uint256 amount
-    )
-        public
-        view
-        override
-        returns (uint256 cashReceived)
-    {
+    function previewRedeem(bytes32 uuid, uint amount) public view override returns (uint cashReceived) {
         if (amount == 0) revert AmountCannotBeZero();
         if (vaultsByUUID[uuid].openedAt == 0) revert InvalidVault();
 
@@ -89,8 +81,8 @@ contract CollarVaultManager is ERC6909TokenSupply, Ownable, ICollarVaultManager 
             // if finalized, calculate final redeem value
             // grab collateral asset value @ exact vault expiration time
 
-            uint256 vaultCash = vaultTokenCashSupply[uuid];
-            uint256 tokenSupply = totalSupply[uint256(uuid)];
+            uint vaultCash = vaultTokenCashSupply[uuid];
+            uint tokenSupply = totalSupply[uint(uuid)];
 
             if (vaultCash == 0) {
                 return 0;
@@ -159,21 +151,20 @@ contract CollarVaultManager is ERC6909TokenSupply, Ownable, ICollarVaultManager 
         IERC20(assetData.collateralAsset).transferFrom(user, address(this), assetData.collateralAmount);
 
         // swap collateral for cash & record cash amount
-        uint256 cashReceivedFromSwap = _swap(assetData);
+        uint cashReceivedFromSwap = _swap(assetData);
         vaultsByUUID[uuid].cashAmount = cashReceivedFromSwap;
 
         // calculate exactly how much cash to lock from the liquidity pool
         // this is equal to (callstrikePercent - 100%) * totalCashReceivedFromSwap
         // first, grab the call strike percent from the call strike tick supplied
-        uint256 tickScaleFactor = CollarPool(liquidityOpts.liquidityPool).tickScaleFactor();
-        uint256 callStrikePercentBps =
-            TickCalculations.tickToBps(liquidityOpts.callStrikeTick, tickScaleFactor);
-        uint256 poolLiquidityToLock =
+        uint tickScaleFactor = CollarPool(liquidityOpts.liquidityPool).tickScaleFactor();
+        uint callStrikePercentBps = TickCalculations.tickToBps(liquidityOpts.callStrikeTick, tickScaleFactor);
+        uint poolLiquidityToLock =
             ((callStrikePercentBps - 10_000) * cashReceivedFromSwap * 1e18) / (10_000) / 1e18;
 
         // calculate the initial collateral price from the swap execution fill
         // this is stored as "unit price times 1e18"
-        uint256 initialCollateralPrice = (cashReceivedFromSwap * 1e18) / (assetData.collateralAmount);
+        uint initialCollateralPrice = (cashReceivedFromSwap * 1e18) / (assetData.collateralAmount);
 
         // set Liquidity Pool Stuff
         vaultsByUUID[uuid].liquidityPool = liquidityOpts.liquidityPool;
@@ -188,7 +179,7 @@ contract CollarVaultManager is ERC6909TokenSupply, Ownable, ICollarVaultManager 
         vaultsByUUID[uuid].callStrikeTick = liquidityOpts.callStrikeTick;
 
         // mint vault tokens (equal to the amount of cash received from swap)
-        _mint(user, uint256(uuid), cashReceivedFromSwap);
+        _mint(user, uint(uuid), cashReceivedFromSwap);
 
         // mint liquidity pool tokens
         CollarPool(liquidityOpts.liquidityPool).openPosition(
@@ -228,11 +219,11 @@ contract CollarVaultManager is ERC6909TokenSupply, Ownable, ICollarVaultManager 
         Vault storage vault = vaultsByUUID[uuid];
 
         // grab all price info
-        uint256 startingPrice = vault.initialCollateralPrice;
-        uint256 putStrikePrice = vault.putStrikePrice;
-        uint256 callStrikePrice = vault.callStrikePrice;
+        uint startingPrice = vault.initialCollateralPrice;
+        uint putStrikePrice = vault.putStrikePrice;
+        uint callStrikePrice = vault.callStrikePrice;
 
-        uint256 finalPrice = CollarEngine(engine).getHistoricalAssetPriceViaTWAP(
+        uint finalPrice = CollarEngine(engine).getHistoricalAssetPriceViaTWAP(
             vault.collateralAsset, vault.cashAsset, vault.expiresAt, 15 minutes
         );
 
@@ -261,8 +252,8 @@ contract CollarVaultManager is ERC6909TokenSupply, Ownable, ICollarVaultManager 
         //  - locked vault cash is given to the user to cover collateral appreciation
         //  - locked pool liquidity is partially used to cover some collateral appreciation
 
-        uint256 cashNeededFromPool = 0;
-        uint256 cashToSendToPool = 0;
+        uint cashNeededFromPool = 0;
+        uint cashToSendToPool = 0;
 
         // CASE 1 - all vault cash to liquidity pool
         if (finalPrice <= putStrikePrice) {
@@ -288,7 +279,7 @@ contract CollarVaultManager is ERC6909TokenSupply, Ownable, ICollarVaultManager 
 
             // CASE 4 - proportional vault cash to user
         } else if (putStrikePrice < finalPrice && finalPrice < startingPrice) {
-            uint256 vaultCashToPool = (
+            uint vaultCashToPool = (
                 (vault.lockedVaultCash * (startingPrice - finalPrice) * 1e32)
                     / (startingPrice - putStrikePrice)
             ) / 1e32;
@@ -300,7 +291,7 @@ contract CollarVaultManager is ERC6909TokenSupply, Ownable, ICollarVaultManager 
 
             // CASE 5 - all vault cash to user, proportional locked pool cash to user
         } else if (callStrikePrice > finalPrice && finalPrice > startingPrice) {
-            uint256 poolCashToUser = (
+            uint poolCashToUser = (
                 (vault.lockedPoolCash * (finalPrice - startingPrice) * 1e32)
                     / (callStrikePrice - startingPrice)
             ) / 1e32;
@@ -322,7 +313,7 @@ contract CollarVaultManager is ERC6909TokenSupply, Ownable, ICollarVaultManager 
             revert InvalidState();
         }
 
-        int256 poolProfit = cashToSendToPool > 0 ? int256(cashToSendToPool) : -int256(cashNeededFromPool);
+        int poolProfit = cashToSendToPool > 0 ? int(cashToSendToPool) : -int(cashNeededFromPool);
 
         if (cashToSendToPool > 0) {
             vault.lockedVaultCash -= cashToSendToPool;
@@ -342,7 +333,7 @@ contract CollarVaultManager is ERC6909TokenSupply, Ownable, ICollarVaultManager 
         emit VaultClosed(msg.sender, address(this), uuid);
     }
 
-    function redeem(bytes32 uuid, uint256 amount) external override {
+    function redeem(bytes32 uuid, uint amount) external override {
         // ensure vault exists
         if (vaultsByUUID[uuid].openedAt == 0) {
             revert InvalidVault();
@@ -354,20 +345,20 @@ contract CollarVaultManager is ERC6909TokenSupply, Ownable, ICollarVaultManager 
         }
 
         // calculate cash redeem value
-        uint256 redeemValue = previewRedeem(uuid, amount);
+        uint redeemValue = previewRedeem(uuid, amount);
 
         emit Redemption(msg.sender, uuid, amount, redeemValue);
 
         // redeem to user & burn tokens
-        _burn(msg.sender, uint256(uuid), amount);
+        _burn(msg.sender, uint(uuid), amount);
         IERC20(vaultsByUUID[uuid].cashAsset).transfer(msg.sender, redeemValue);
     }
 
-    function withdraw(bytes32 uuid, uint256 amount) public override {
+    function withdraw(bytes32 uuid, uint amount) public override {
         if (msg.sender != user) revert NotCollarVaultOwner();
         if (vaultsByUUID[uuid].openedAt == 0) revert InvalidVault();
 
-        uint256 loanBalance = vaultsByUUID[uuid].loanBalance;
+        uint loanBalance = vaultsByUUID[uuid].loanBalance;
 
         // withdraw from user's loan balance
         if (amount > loanBalance) {
@@ -438,7 +429,7 @@ contract CollarVaultManager is ERC6909TokenSupply, Ownable, ICollarVaultManager 
         }
     }
 
-    function _swap(AssetSpecifiers calldata assets) internal returns (uint256 cashReceived) {
+    function _swap(AssetSpecifiers calldata assets) internal returns (uint cashReceived) {
         // approve the dex router so we can swap the collateral to cash
         IERC20(assets.collateralAsset).approve(CollarEngine(engine).dexRouter(), assets.collateralAmount);
 
@@ -461,12 +452,12 @@ contract CollarVaultManager is ERC6909TokenSupply, Ownable, ICollarVaultManager 
         }
     }
 
-    function _mint(address account, uint256 id, uint256 amount) internal {
+    function _mint(address account, uint id, uint amount) internal {
         balanceOf[account][id] += amount;
         totalSupply[id] += amount;
     }
 
-    function _burn(address account, uint256 id, uint256 amount) internal {
+    function _burn(address account, uint id, uint amount) internal {
         balanceOf[account][id] -= amount;
         totalSupply[id] -= amount;
     }
