+++ test/utils/TestERC20.sol
@@ -12,7 +12,7 @@ import { ERC20 } from "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 contract TestERC20 is ERC20 {
     constructor(string memory name_, string memory symbol_) ERC20(name_, symbol_) { }
 
-    function mint(address to, uint256 amount) external {
+    function mint(address to, uint amount) external {
         _mint(to, amount);
     }
 }
